post

+++ app/Http/Controllers/TasksController.php
@@ -14,11 +14,20 @@ class TasksController extends Controller
    
     public function index()
     {
-        $tasks= Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('users.show', $data);
+        }else {
+            return view('welcome');
+        }
     }
 
    
@@ -43,6 +52,10 @@ class TasksController extends Controller
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
+        
+        $request->user()->microposts()->create([
+            'content' => $request->content,
+        ]);
 
         return redirect('/');
     }
@@ -82,9 +95,9 @@ class TasksController extends Controller
 
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
-
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
         return redirect('/');
     }
 }
